requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
_pytorch_select=0.2=gpu_0
addict=2.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
async-generator=1.10=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=22.1.0=pypi_0
autopep8=1.5.3=py_0
babel=2.11.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=pypi_0
bel-resources=0.0.3=pypi_0
bioregistry=0.6.30=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cairo=1.14.12=h8948797_3
certifi=2022.12.7=py37h06a4308_0
cffi=1.13.2=py37h8022711_0
chardet=5.1.0=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
click-plugins=1.1.1=pypi_0
cloudpickle=1.3.0=py_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
curies=0.4.0=pypi_0
cycler=0.10.0=py_2
cython=0.29.33=pypi_0
cytoolz=0.10.1=py37h516909a_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
dask-core=2.11.0=py_0
debugpy=1.6.4=pypi_0
decorator=4.4.1=py_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.13=pypi_0
entrypoints=0.4=pypi_0
exceptiongroup=1.1.0=pypi_0
fastjsonschema=2.17.1=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.0=he983fc9_1
frozenlist=1.3.3=pypi_0
glib=2.63.1=h5a9c865_0
greenlet=2.0.1=pypi_0
gudhi=3.7.1=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
h11=0.14.0=pypi_0
h5py=2.10.0=nompi_py37h513d04c_102
hdbscan=0.8.24=py37hc1659b7_0
hdf5=1.10.5=nompi_h3c11f04_1104
hopcroftkarp=1.2.5=pypi_0
humanize=4.4.0=pypi_0
icu=58.2=he6710b0_3
idna=3.4=pypi_0
imageio=2.8.0=py_0
imbalanced-learn=0.9.0=pypi_0
imblearn=0.0=pypi_0
importlib-metadata=5.1.0=pypi_0
importlib-resources=5.10.0=pypi_0
ipykernel=6.16.2=pypi_0
ipynb=0.5.1=pypi_0
ipython=7.34.0=pypi_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.0.6=pypi_0
jedi=0.17.0=py37_0
jinja2=3.1.2=pypi_0
joblib=0.14.1=py_0
jpeg=9c=h14c3975_1001
jsanimation=0.1=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.4.7=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.24.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.7=pypi_0
kaggle=1.5.12=pypi_0
kiwisolver=1.1.0=py37hc9558a2_0
labelimg=1.8.6=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_8
libffi=3.2.1=he1b5a44_1006
libgcc-ng=9.2.0=h24d8f2e_2
libgfortran-ng=7.3.0=hdf63c60_5
libllvm11=11.1.0=h3826bc1_1
libpng=1.6.37=hed695b0_0
libprotobuf=3.11.4=h8b12597_0
libstdcxx-ng=9.2.0=hdf63c60_2
libtiff=4.1.0=hc3755c2_3
libuuid=1.0.3=h7f8727e_2
libxcb=1.15=h7f8727e_0
libxml2=2.9.9=hea5a465_1
llvm-openmp=9.0.1=hc9558a2_2
llvmlite=0.38.0=py37h4ff587b_0
lxml=4.9.2=pypi_0
lz4-c=1.8.3=he1b5a44_1001
markupsafe=2.1.2=pypi_0
matplotlib-base=3.2.2=py37hef1b27d_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
mkl=2019.5=281
mkl-service=2.3.0=py37h516909a_0
mkl_fft=1.1.0=py37hc1659b7_1
mkl_random=1.1.0=py37hb3f55d8_0
more-click=0.1.1=pypi_0
more-itertools=9.0.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
multisplitby=0.0.1=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.4.0=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.1=hf484d3e_1002
nest-asyncio=1.5.6=pypi_0
networkx=2.4=py_0
nibabel=3.2.1=pyhd8ed1ab_0
ninja=1.10.0=hc9558a2_0
notebook=6.5.4=pypi_0
notebook-shim=0.2.3=pypi_0
numba=0.55.1=py37h51133e4_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
olefile=0.46=py_0
openai=0.26.5=pypi_0
openbabel=2.4.1=py37_6
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1t=h7f8727e_0
outcome=1.2.0=pypi_0
packaging=20.9=pyh44b312d_0
pandas=1.0.1=py37h0573a6f_0
pandocfilters=1.5.0=pypi_0
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
persim=0.3.1=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickle5=0.0.12=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=6.2.1=py37hd70f55b_1
pip=23.2.1=pypi_0
pixman=0.40.0=h7f8727e_1
pkgutil-resolve-name=1.3.10=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.11.4=py37he1b5a44_0
psutil=5.9.4=pypi_0
psycopg2-binary=2.9.5=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pybel=0.15.5=pypi_0
pycocotools=2.0=pypi_0
pycodestyle=2.6.0=py_0
pycparser=2.19=py_2
pydantic=1.10.2=pypi_0
pydicom=2.1.2=pyhd3deb0d_0
pydot=1.4.2=pypi_0
pygments=2.8.1=pyhd3eb1b0_0
pyparsing=2.4.6=py_0
pyqt5=5.15.9=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.12.1=pypi_0
pyrsistent=0.19.2=pypi_0
pysocks=1.7.1=pypi_0
pystow=0.4.7=pypi_0
python=3.7.6=h357f687_2
python-dateutil=2.8.2=pypi_0
python-graphviz=0.20.1=pypi_0
python-slugify=8.0.0=pypi_0
python_abi=3.7=1_cp37m
pytorch=1.3.1=cuda100py37h53c1284_0
pytorch-3dunet=1.2.4=py37_0
pytrie=0.4.0=pypi_0
pytz=2019.3=py_0
pywavelets=1.1.1=py37hc1659b7_0
pyyaml=5.3=py37h516909a_0
pyzmq=24.0.1=pypi_0
qtconsole=5.4.3=pypi_0
qtpy=2.3.1=pypi_0
ratelimit=2.2.1=pypi_0
rdkit=2022.9.2=pypi_0
readline=8.0=hf8c457e_0
requests=2.28.1=pypi_0
requests-file=1.5.1=pypi_0
retrying=1.3.4=pypi_0
revchatgpt=2.0.3=pypi_0
ripser=0.6.4=pypi_0
scikit-image=0.16.2=py37hb3f55d8_0
scikit-learn=1.0.2=pypi_0
scipy=1.4.1=py37h0b6359f_0
seaborn=0.12.1=pypi_0
selenium=4.8.0=pypi_0
send2trash=1.8.2=pypi_0
setuptools=45.2.0=py37_0
simpleitk=2.2.0=pypi_0
six=1.14.0=py37_0
sklearn=0.0.post1=pypi_0
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlalchemy=1.4.44=pypi_0
sqlite=3.30.1=hcee41ef_0
sympy=1.10.1=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.2=pypi_0
tensorboardx=2.0=py_0
terminado=0.17.1=pypi_0
text-unidecode=1.3=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.10=hed695b0_0
tls-client=0.1.8=pypi_0
toml=0.10.1=py_0
toolz=0.10.0=py_0
torchsummary=1.5.1=pypi_0
torchvision=0.4.2=cuda100py37hecfc37a_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.6.0=pypi_0
trio=0.22.0=pypi_0
trio-websocket=0.9.2=pypi_0
typing-extensions=4.4.0=pypi_0
undetected-chromedriver=3.4.6=pypi_0
urllib3=1.26.12=pypi_0
vtk=9.2.6=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.2=pypi_0
websockets=10.4=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.34.2=py_1
widgetsnbextension=4.0.7=pypi_0
wrapt=1.14.1=pypi_0
wsproto=1.2.0=pypi_0
xlrd=1.2.0=py37_0
xz=5.2.4=h14c3975_1001
yaml=0.2.2=h516909a_1
yarl=1.8.2=pypi_0
zipp=3.11.0=pypi_0
zlib=1.2.11=h516909a_1006
zstd=1.4.4=h3b9ef0a_1
